book_api_project/requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1
pytz==2025.1
setuptools==77.0.1
sqlparse==0.5.3